+++ include/gdk.h
@@ -176,6 +176,17 @@ GDK_API int GA_get_proxy_settings(struct GA_session* session, GA_json** output);
 GDK_API int GA_get_wallet_identifier(const GA_json* net_params, const GA_json* params, GA_json** output);
 
 /**
+ * Operate on cached session data.
+ *
+ * :param session: The session to use.
+ * :param details: The :ref:`cache-control-request` giving the operation to perform.
+ * :param call: Destination for the resulting GA_auth_handler to complete the action.
+ *|     The call handlers result is :ref:`cache-control-result`.
+ *|     Returned GA_auth_handler should be freed using `GA_destroy_auth_handler`.
+ */
+GDK_API int GA_cache_control(struct GA_session* session, GA_json* details, struct GA_auth_handler** call);
+
+/**
  * Make a request to an http server.
  *
  * :param session: The session to use.